pyproject.toml
[project]
name = "varsdaa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.8",
    "django-allauth[socialaccount]>=65.13.1",
    "iommi>=7.20.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "pytest-django>=4.11.1",
    "pytest-sugar>=1.1.1",
    "ruff>=0.14.6",
]
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
varsdaa = "varsdaa"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "varsdaa.test.settings"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
ignore = [
]
select = [
    "E",
    "F",
    "I",
    "N",
    "W",
    "UP",
]

[tool.ruff.format]
quote-style = "preserve"